% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_rank_plot.R
\name{distance_rank_plot}
\alias{distance_rank_plot}
\title{Pairwise Document Distance Rank Plot}
\usage{
distance_rank_plot(distance_rank_test_results, display_raw_rankings = FALSE,
  remove_labels = FALSE, num_docs = NULL, text_size = 1)
}
\arguments{
\item{distance_rank_test_results}{The ouptu from the distance_rank_test()
function.}

\item{display_raw_rankings}{Logical indicating whether raw ranking
differences should be displayed (as opposed to relative differences).}

\item{num_docs}{If display_raw_rankings = TRUE, the number of documents in
the corpus.}

\item{text_size}{The cex for text in dot plot generated by function.}

\item{remove_label}{Option to remove preprocessing step labels. Defaults to
FALSE.}
}
\value{
A plot
}
\description{
Document pair distance ranking plots.
}

