% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_rank_test.R
\name{distance_rank_test}
\alias{distance_rank_test}
\title{Pairwise Document Distance Rank Test}
\usage{
distance_rank_test(distance_matrices, choices, labels = NULL,
  method = c("top", "distribution", "continuous"), baseline_index = 128,
  text_size = 1, num_comparisons = 300)
}
\arguments{
\item{distance_matrices}{A list of document distance matrices.}

\item{choices}{A dataframe indicating whether a preprocessing step was
applied or not, for each preprocessing step.}

\item{labels}{Optional argument giving names for each preprocessing step.}

\item{method}{These are different ways of looking for unusual preprocessing
steps. If "continuous" is selected, then the average absolute difference in
distances between documents is used.}

\item{baseline_index}{The index of the baseline distance matrix against which
we are comparing. Defautls to 128, which is the most minimal preprocessing
for our current implementation.}

\item{text_size}{The cex for text in dot plot generated by function.}

\item{num_comparisons}{If method = "distribution", the number of ranks to use
in calculating average difference. Defautls to 300.}
}
\value{
A result list object.
}
\description{
Document pair distance ranking test metrics.
}

