% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_preprocessing.R
\name{factorial_combinations}
\alias{factorial_combinations}
\title{Create factorial combinations of pre-processing choices.}
\usage{
factorial_combinations(choices)
}
\arguments{
\item{choices}{an ordered and named list of logical vectors which indicator
which pre-processing parameters to include in the factorial combination of
choices. The list must include the following entries (in order):
removePunctuation, removeNumbers, lowercase, stem, removeStopwords,
infrequent_terms, use_ngrams.}
}
\description{
Checks user input or supplies default values
}
