% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_preprocessing.R
\name{factorial_preprocessing}
\alias{factorial_preprocessing}
\title{A function to perform factorial preprocessing of a corpus of texts
into document-frequency matrices.}
\usage{
factorial_preprocessing(text)
}
\arguments{
\item{text}{A vector of strings (one per document) or Quanteda corpus object
from which we wish to form a document-term matrix.}
}
\value{
A list object containing permutations of the document-term matrix.
}
\description{
Preprocesses a corpus of texts into a document-frequency matrix
in thirty-two different ways.
}

