% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel_comparison.R
\name{mantel_comparison}
\alias{mantel_comparison}
\title{Ensemble Mantel Tests}
\usage{
mantel_comparison(distance_matrices, labels = NULL, permutations = 1000)
}
\arguments{
\item{distance_matrices}{A list of document distance matrices generated by
the `scaling_comparison()` and returned in the `$distance_matrices` field.}

\item{labels}{Optional argument giving names for each preprocessing step.
This is generated by the `factorial_preprocessing()` function and returned in
the `$labels` field.}

\item{permutations}{The number of permutations to be used in each Mantel
test. Defaults to 1000.}
}
\value{
A result list object where the first entry is a matrix summarizing
mantel test statistics. The second object in the list is a matrix of the
values described above. The third object is a list of all raw mantel results.
}
\description{
Calculates Mantel test statistics for differences between
distance matrices for a list of distance matrices (one per preprocessing
method) supplied by the `scaling_comparison()` function.
}

