% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel_comparison_to_base.R
\name{mantel_comparison_to_base}
\alias{mantel_comparison_to_base}
\title{Ensemble Mantel Tests}
\usage{
mantel_comparison_to_base(distance_matrices, names = NULL,
  permutations = 1000, base_dfm_index = 128, text_size = 1,
  return_values = FALSE)
}
\arguments{
\item{distance_matrices}{A list of document distance matrices.}

\item{names}{Optional argument giving names for each preprocessing step.}

\item{permutations}{the number of permutations to be used in each Mantel
test. Defaults to 1000.}

\item{base_dfm_index}{Which dfm should be used as a base case for comparing
r statistics with bootstrapped confidence intervals.}

\item{text_size}{THe cex for the x labels, defaults to 1.}

\item{return_values}{Logical indicating whether test statistics and
confidencebounds should be returned as a data.frame or not. Defaults to FALSE.}
}
\value{
A data.frame with mantel statistics and 95 percent confidence
intervals comparing all other preprocessing choices to base case and or a
plot of confidence intervals.
}
\description{
Calculates Mantel test statistics for differences between
distance matrices for a list of distance matrices (one per preprocessing
method) supplied by the scaling_comparison() function and a base case.
}

