% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_k_comparison.R
\name{optimal_k_comparison}
\alias{optimal_k_comparison}
\title{Optimal Topic Model k Comparison}
\usage{
optimal_k_comparison(dfm_object_list, interval = c(2, 500), names = NULL,
  parallel = FALSE, cores = 1)
}
\arguments{
\item{dfm_object_list}{A list of quanteda dfm() objects.}

\item{interval}{The maximum and minimum number of topics to be searched over
defaults to 2 for minimum and 500 for maximum.}

\item{names}{optinal names for each dfm to make downstream interpretation
easier. Defaults to NULL.}

\item{parallel}{Logical indicating whether model fitting should be
performed in parallel. Defualts to FALSE.}

\item{cores}{Defualts to 1, can be set to any number less than or equal to
the number of cores on one's computer.}
}
\value{
A vector containing the optiaml k for each dfm.
}
\description{
Calculate the optimal number of topics for LDA using perplexity
for each dfm.
}

