% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_choice_regression.R
\name{preprocessing_choice_regression}
\alias{preprocessing_choice_regression}
\title{Preprocessing Choice Regressions}
\usage{
preprocessing_choice_regression(Y, choices, dataset = "UK",
  base_case_index = 128)
}
\arguments{
\item{Y}{A vector of length 128 (usually) containing a numeric outcome
variable.}

\item{choices}{A 128 x 7 data.frame produced by the factorial_preprocessing
function.}

\item{dataset}{The name to be given to the data we are analyzing.}

\item{base_case_index}{An optional argument whic removes a base case row from
the choices data before performing the regrerssion.}
}
\value{
A data.frame
}
\description{
Assessing the effects of preprocessing decisions on an outcome
variable.
}

