% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Documentation.R
\docType{package}
\name{preptest}
\alias{preptest}
\alias{preptest-package}
\title{preptest: fucntions to asses the effects of text preprocessing decisions}
\description{
preptest: fucntions to asses the effects of text preprocessing decisions
}
\section{preptest functions}{

To use this package, You will first want to check out the factorial_preprocessing() function which will take raw data and transform it into document-frequency matrices using a factorial desig and five different preprocessing decisions.
}

