% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Documentation.R, R/preText.R
\docType{package}
\name{preText}
\alias{preText}
\alias{preText-package}
\title{preText: fucntions to asses the effects of text preprocessing decisions}
\usage{
preText(preprocessed_documents, dataset_name = "Documents",
  distance_method = "cosine", num_comparisons = 50, parallel = FALSE,
  cores = 1, verbose = TRUE)
}
\arguments{
\item{preprocessed_documents}{A list object generated by the
`factorial_preprocessing()` function.}

\item{dataset_name}{A string indicating the name to be associated with the
results. Defaults to "Documents".}

\item{distance_method}{The method that should be used for calculating
document distances. Defualts to "cosine".}

\item{num_comparisons}{If method = "distribution", the number of ranks to use
in calculating average difference. Defautls to 50.}

\item{parallel}{Logical indicating whether factorial prerpocessing should be
performed in parallel. Defualts to FALSE.}

\item{cores}{Defualts to 1, can be set to any number less than or equal to
the number of cores on one's computer}

\item{verbose}{Logical indicating whether more information should be printed
to the screen to let the user know about progress. Defaults to TRUE.}
}
\value{
A result list object.
}
\description{
preText: fucntions to asses the effects of text preprocessing decisions

Calculates preText scores for each preprocessing specification.
}
\section{preText functions}{

To use this package, You will first want to check out the factorial_preprocessing() function which will take raw data and transform it into document-frequency matrices using a factorial desig and five different preprocessing decisions.
}

