% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preText_score_plot.R
\name{preText_score_plot}
\alias{preText_score_plot}
\title{preText specification plot}
\usage{
preText_score_plot(preText_results, display_raw_rankings = FALSE,
  remove_labels = FALSE, num_docs = NULL, text_size = 1)
}
\arguments{
\item{preText_results}{The ouput from the `preText_test()` or
`preText()` functions.}

\item{display_raw_rankings}{Logical indicating whether raw ranking
differences should be displayed (as opposed to relative differences).}

\item{remove_labels}{Option to remove preprocessing step labels. Defaults to
FALSE.}

\item{num_docs}{If display_raw_rankings = TRUE, the number of documents in
the corpus.}

\item{text_size}{The cex for text in dot plot generated by function.}
}
\value{
A plot
}
\description{
preText plots for each preprocessing specification.
}

