% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_coefficient_plot.R
\name{regression_coefficient_plot}
\alias{regression_coefficient_plot}
\title{Regression Coefficient Plot}
\usage{
regression_coefficient_plot(data, text_size = 1, remove_intercept = FALSE,
  title = "")
}
\arguments{
\item{data}{A data.frame produced by the `preprocessing_choice_regression`
function.}

\item{text_size}{The size of the text to be displayed. Defaults to 1.}

\item{remove_intercept}{Logical indicating whwter intercept coefficiet sould
be plotted. Defalts to FALSE.}

\item{title}{The title the user wishes to give the lot, whic will be diplayed
instead of the axis title.}
}
\value{
A plot
}
\description{
Easy plotting for multiple regression results on same axes.
}

