% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_filter.R
\name{temporal_filter}
\alias{temporal_filter}
\title{Remove terms that only appear in a less than a specified proportion of
years.}
\usage{
temporal_filter(dfm_object, years, proportion_threshold = 0.5)
}
\arguments{
\item{dfm_object}{A quanteda dfm() object.}

\item{years}{A numeric vector giving the year for each document}

\item{proportion_threshold}{proportion of years a term must be included in
to be included in the dfm.}
}
\value{
A reduced dfm.
}
\description{
Removes terms appearing in less than a specifiec percentage of
years in the data.
}

