% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_key_term_plot.R
\name{topic_key_term_plot}
\alias{topic_key_term_plot}
\title{Plot Prevalence of Topiuc Key Terms}
\usage{
topic_key_term_plot(topic_key_term_results, key_term_columns = 2:6,
  custom_col_names = c("Iraq", "Terrorism", "Al Qaeda", "Insurance",
  "Stem Cell"), custom_labels = c("0\%", "<1\%", "1-2\%", "2-3\%", "3-4\%",
  "4-5\%", "5-6\%", "6-7\%", "7-8\%", "8-9\%", "9-10\%", "10\%+"),
  one_matrix = FALSE, thresholds = c(-1e-04, 0, 0.0099, 0.0199, 0.0299,
  0.0399, 0.0499, 0.0599, 0.0699, 0.0799, 0.0899, 0.0999),
  labs = labels[65:128], heat_ramp = FALSE)
}
\arguments{
\item{topic_key_term_results}{A data.frame.}

\item{key_term_columns}{The columns containing key term results.}

\item{custom_col_names}{Names for the key terms.}

\item{custom_labels}{Labels for the provided key. Must be of length 12.}

\item{one_matrix}{Logical indicating whether results shoudl be displayed as
a one column matrix. Defaults to FALSE.}

\item{thresholds}{A numeric vector of length 11 with threshold for inclusion
in various heat map categroies.}

\item{labs}{Labels for the preprocessing choices being used in the analysis.}

\item{heat_ramp}{Option to use heat ramp (yellow-red-purple) instead of a
white tp blue ramp.}
}
\value{
A plot
}
\description{
Plotting of key terms across preprocessing decisions.
}

