% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_novelty_score.R
\name{topic_novelty_score}
\alias{topic_novelty_score}
\title{Topic Top-Terms Novelty Score}
\usage{
topic_novelty_score(top_terms_matrix, row_range = NULL)
}
\arguments{
\item{top_terms_matrix}{A character matrix or data.frame containing top
terms for all topics.}

\item{row_range}{Optional argument specifying range of rows to keep. This is
useful if we are only interested in the top 10 or 20 terms, but we have a
matrix with the top 50 terms for each topic.}
}
\value{
A novelty score.
}
\description{
Calculates the novelty score for a character matrix displaying
topic top-terms for all topics.
}

