% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{UK_Manifestos}
\alias{UK_Manifestos}
\title{Full text of 69 UK party manifestos from 1918-2001.}
\format{A character vector of length 69, with one document per entry.}
\source{
Aggregated from several online resources but see, for example:
\url{http://www.politicsresources.net/area/uk/man.htm}
}
\usage{
UK_Manifestos
}
\description{
A dataset of party manifestos for Liberal, Labour, and Conservative parties
for every general election in the UK between 1918 and 2001 (23 elections).
}
\keyword{datasets}

