% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prediction_errors.R
\name{calculate_prediction_errors}
\alias{calculate_prediction_errors}
\title{Calculate Mean Prediction Error for Preprocessing Decisions}
\usage{
calculate_prediction_errors(positions_list, preprocessing_choices)
}
\arguments{
\item{positions_list}{A list of scaled document positions.}

\item{preprocessing_choices}{A data frame containing binary indicators of
wether each preprocessing decision was applied for each dfm.
red.}
}
\value{
A vector of mean prediction errors.
}
\description{
Use scaled positions to predict preprocessing decisions.
}

