% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_scaling_test.R
\name{dfm_scaling_test}
\alias{dfm_scaling_test}
\title{comparison of dfms using scaling, with a test for difference from the
mean dfm.}
\usage{
dfm_scaling_test(scaling_results, labels, dimensions = 2,
  distance_method = "cosine", method = c("distances", "positions"),
  return_positions = FALSE)
}
\arguments{
\item{scaling_results}{A list object produced by the `scaling_comparison()`
function.}

\item{labels}{A character vector with labels for each dfm, based on
preprocessing choices.}

\item{dimensions}{The number of dimensions to be used by the multidimensional
scaling algorithm. Defualts to 2.}

\item{distance_method}{The method that should be used for calculating
document distances. Defualts to "cosine".}

\item{method}{Should the raw distances or scaled document positions be used
for scaling? can be one of c("distances","positions"), defaults to
"distances".}

\item{return_positions}{Logical indicating whether dfm positions should be
returned as a data.frame. Defaults to FALSE}
}
\value{
A result list object or a plot, or both.
}
\description{
Scale each dfm into a 2-d space and test for outliers.
}

