% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_preprocessing.R
\name{factorial_preprocessing}
\alias{factorial_preprocessing}
\title{A function to perform factorial preprocessing of a corpus of texts
into document-frequency matrices.}
\usage{
factorial_preprocessing(text, use_ngrams = TRUE,
  filter_infrequent_terms = TRUE, infrequent_term_threshold = 0.01,
  parallel = FALSE, cores = 1, intermediate_directory = NULL,
  parameterization_range = NULL, return_results = TRUE)
}
\arguments{
\item{text}{A vector of strings (one per document) or Quanteda corpus object
from which we wish to form a document-term matrix.}

\item{use_ngrams}{Option to extract 1,2, and 3-grams from the text as another
potential preprocessing step. Defualts to TRUE.}

\item{filter_infrequent_terms}{Option to remove all terms appearing in less
than 1 percent of documents as a preprocessing decision. Defaults to TRUE.}

\item{infrequent_term_threshold}{A proportion threshold at which infrequent
terms are to be filtered. Defaults to 0.01 (terms that appear in less than
1 percent of documents).}

\item{parallel}{Logical indicating whether factorial prerpocessing should be
performed in parallel. Defualts to FALSE.}

\item{cores}{Defualts to 1, can be set to any number less than or equal to
the number of cores on one's computer.}

\item{intermediate_directory}{Optional path to a directory where each dfm
will be saved as an intermediate step. The file names will follow the
convention intermediate_dfm_i.Rdata, where i is the index of the combination
of preprocessing choices. The function will then attempt to read all of the
dfm's back into a list if return_results = TRUE (by default), or simply end
the function call if return_results = FALSE. This can be a useful option if
the user is preprocessing a corpus that would make a dfm list that was
impractical to work with due to its size.}

\item{parameterization_range}{Defaults to NULL, but can be set to a numeric
vector of indexes relating to preprocessing decisions. This can be used to
restart large analyses after power failure.}

\item{return_results}{Defaults to TRUE, can be set to FALSE to prevent an
overly large dfm list from being created.}
}
\value{
A list object containing permutations of the document-term matrix.
}
\description{
Preprocesses a corpus of texts into a document-frequency matrix
in thirty-two different ways.
}

