% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel_comparison.R
\name{mantel_comparison}
\alias{mantel_comparison}
\title{Ensemble Mantel Tests}
\usage{
mantel_comparison(distance_matrices, names = NULL, permutations = 999)
}
\arguments{
\item{distance_matrices}{A list of document similarity matrices.}

\item{names}{Optional argument giving names for each preprocessing step.}

\item{permutations}{the number of permutations to be used in each Mantel
test. Defaults to 999.}
}
\value{
A result list object where the first entry is a matrix summarizing
mantel test statististics. The fourth column of this matrix records the
difference between the test statistic and the 99th percentile of the null
distribution, and is preferred for comparing between runs. A positive value
indicates significance at atleast the 0.01 level, while a negative value
indciates insignificance. The second object in the list is a matrix of the
values described above. The third object is a list of all raw mantel results.
}
\description{
Calculates Mantel test statistics for differences between
distance matrices for a list of distance matrices (one per preprocessing
method) supplied by the scaling_comparison() function.
}

