% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel_comparison.R
\name{mantel_comparison}
\alias{mantel_comparison}
\title{Ensemble Mantel Tests}
\usage{
mantel_comparison(distance_matrices, names = NULL, permutations = 1000)
}
\arguments{
\item{distance_matrices}{A list of document distance matrices.}

\item{names}{Optional argument giving names for each preprocessing step.}

\item{permutations}{the number of permutations to be used in each Mantel
test. Defaults to 1000.}
}
\value{
A result list object where the first entry is a matrix summarizing
mantel test statististics. The second object in the list is a matrix of the
values described above. The third object is a list of all raw mantel results.
}
\description{
Calculates Mantel test statistics for differences between
distance matrices for a list of distance matrices (one per preprocessing
method) supplied by the scaling_comparison() function.
}

