% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_comparison.R
\name{scaling_comparison}
\alias{scaling_comparison}
\title{Scaling Comparison.}
\usage{
scaling_comparison(dfm_object_list, dimensions = 2,
  similarity_method = "cosine")
}
\arguments{
\item{dfm_object_list}{A list of quanteda dfm() objects.}

\item{dimensions}{THe number of dimensions to be used by the multidimensional
scaling algorithm. Defualts to 2.}

\item{similarity_method}{The method that should be used for calculating
document similarities. Defualts to "cosine".}
}
\value{
A result list object.
}
\description{
Scale each dfm and return a list of similarity matrices and
scaled document positions.
}

