% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfish_comparison.R
\name{wordfish_comparison}
\alias{wordfish_comparison}
\title{Wordfish Comparison.}
\usage{
wordfish_comparison(dfm_object_list, years, anchors = c(1, 24),
  proportion_threshold = 1, document_inidices = NULL)
}
\arguments{
\item{dfm_object_list}{A list of quanteda dfm() objects.}

\item{years}{A numeric vector giving the year for each document.}

\item{anchors}{A numeric vetor of length two used to anchor the wordfish
estimates. Defaults to c(1,24) which should work for the UK parlaiment docs.}

\item{proportion_threshold}{proportion of years a term must be included in
to be included in the Wordfish analysis.}

\item{document_inidices}{An option vector of row indicies to be used. Useful
for using a subset of the data for analysis.}
}
\value{
A result list object
}
\description{
Calculated Wordfish scores for a list of dfm objects with
temporal filtering.
}

