% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfish_rank_plot.R
\name{wordfish_rank_plot}
\alias{wordfish_rank_plot}
\title{Plot of Wrodfish rankings of documents}
\usage{
wordfish_rank_plot(wordfish_results, labels, invert = TRUE,
  ranking = c("Lab1983", "Lab1987", "Lab1992", "Lab1997", "Con1997",
  "Con1992", "Con1987", "Con1983"), black_white = FALSE)
}
\arguments{
\item{wordfish_results}{The output from the `wordfish_comparison()` function.}

\item{labels}{A character vector giving the names for each preprocessing step.}

\item{invert}{Logical indicating whether wordfish score rankings should be
reversed internally.}

\item{ranking}{A character vector containing the correctly ranked document
names.}

\item{black_white}{Logical, defualts to FALSE. If FALSE then results are
displayed on a red-blue scale. IF TRUE, then mis-ordered documents are
colored black.}
}
\value{
A plot.
}
\description{
Coloration by ground-truth ranking.
}

